+++ 유창민/프레임워크 3.3/Scene.h
@@ -80,15 +80,15 @@ public:
 	CGameObject* SelectBar = NULL; //스킬선택
 private:
 	//여기다 물체 배치에 대한 상수들 넣는다.
-	float BigWall_X1 = 300 * sinf(0.4f * MMPE_PI);
-	float BigWall_Z1 = (600 * cosf(0.4f * MMPE_PI) + 600) / 2;
+	float BigWall_X1 = 300.0f * sinf(0.4f * (float)MMPE_PI);
+	float BigWall_Z1 = (600.0f * cosf(0.4f * (float)MMPE_PI) + 600.0f) / 2.0f;
 
-	float BigWall_X2 = ((600 * sinf(0.4f * MMPE_PI)) + (600 * sinf(0.8f * MMPE_PI))) / 2;
-	float BigWall_Z2 = ((600 * cosf(0.4f * MMPE_PI)) + (-600 * cosf(0.2f * MMPE_PI))) / 2;
+	float BigWall_X2 = ((600.0f * sinf(0.4f * (float)MMPE_PI)) + (600.0f * sinf(0.8f * (float)MMPE_PI))) / 2.0f;
+	float BigWall_Z2 = ((600.0f * cosf(0.4f * (float)MMPE_PI)) + (-600 * cosf(0.2f * (float)MMPE_PI))) / 2.0f;
 
-	float BigWall_Z3 = -600 * cosf(0.2f * MMPE_PI);
+	float BigWall_Z3 = -600.0f * cosf(0.2f * (float)MMPE_PI);
 
-	float BigWall_Rad1 = MMPE_PI / 5; //36degree
+	float BigWall_Rad1 = (float)(MMPE_PI / 5.0f); //36degree
 
 
 };
